requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
distlib==0.4.0
Django==5.2.5
filelock==3.19.1
frozenlist==1.7.0
g4f==0.5.7.6
gunicorn==23.0.0
idna==3.10
multidict==6.6.3
nest-asyncio==1.6.0
numpu==0.1.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==12.0.0
platformdirs==4.4.0
propcache==0.3.2
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stevedore==5.5.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
xlsxwriter==3.2.9
yarl==1.20.1
